requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
pyyaml>=6.0.1
redis>=5.0.0
celery>=5.3.0
ultralytics>=8.0.0
torch>=2.1.0
torchvision>=0.16.0
